Refactor: move versioning-related QR-helpers into "versioning" package
